use templater to build injection script
