+++ .github/actions/build-ab/contents/entitlement-check.js
@@ -1,0 +1,45 @@
+try {
+    var xhr = new XMLHttpRequest()
+    xhr.open('POST', '/graphql')
+    xhr.setRequestHeader('authority', '${entitlementsEndpoints[environment]}')
+    xhr.setRequestHeader('cache-control','no-cache')
+    xhr.setRequestHeader('content-type', 'application/json; charset=utf-8')
+    xhr.setRequestHeader('newrelic-requesting-services','platform|nr1-ui')
+    xhr.setRequestHeader('pragma','no-cache')
+
+    xhr.send(
+        JSON.stringify({
+            'query': 'query PlatformEntitlementsQuery($names: [String]!) { currentUser { id authorizedAccounts { id entitlements(filter: {names: $names}) { name __typename } __typename } __typename } }',
+            'variables': {
+                'names': [
+                    'hipaa',
+                    'fedramp'
+                ]
+            }
+        })
+    )
+
+    xhr.addEventListener('load', function(evt){
+      try{
+        var entitlementsData = JSON.parse(evt.currentTarget.responseText)
+        // call the newrelic api(s) after evaluating entitlements...
+        var canRunSr = true
+        if (entitlementsData.data && entitlementsData.data.currentUser && entitlementsData.data.currentUser.authorizedAccounts && entitlementsData.data.currentUser.authorizedAccounts.length) {
+          for (var i = 0; i < entitlementsData.data.currentUser.authorizedAccounts.length; i++){
+            var ent = entitlementsData.data.currentUser.authorizedAccounts[i]
+            if (!ent.entitlements || !ent.entitlements.length) continue
+            for (var j = 0; j < ent.entitlements.length; j++){
+              if (ent.entitlements[j].name === 'fedramp' || ent.entitlements[j].name === 'hipaa') canRunSr = false
+            }
+          } 
+        }
+        if (canRunSr && newrelic && !!newrelic.start) newrelic.start('session_replay')
+      } catch(e){
+        // something went wrong...
+        if (!!newrelic && !!newrelic.noticeError) newrelic.noticeError(e)
+      }
+    })
+  } catch(err){
+    // we failed our mission....
+    if (!!newrelic && !!newrelic.noticeError) newrelic.noticeError(err)
+  }
