+++ .github/actions/build-ab/package.json
@@ -4,9 +4,10 @@
   "private": true,
   "type": "module",
   "devDependencies": {
-    "@newrelic/browser-agent.actions.shared-utils": "file:../shared-utils",
     "@aws-sdk/client-s3": "^3.385.0",
     "@aws-sdk/client-sts": "^3.385.0",
+    "@newrelic/browser-agent.actions.shared-utils": "file:../shared-utils",
+    "handlebars": "^4.7.8",
     "node-fetch": "^3.3.2",
     "uuid": "^9.0.0",
     "yargs": "^17.7.2"
